Breaking Change: Move all APIGateway attributes to a their own namespace Amazon.Lambda.Annotations.APIGateway
