+++ Libraries/src/Amazon.Lambda.Annotations.SourceGenerator/Models/Attributes/FromQueryAttributeBuilder.cs
@@ -1,3 +1,4 @@
+using Amazon.Lambda.Annotations.APIGateway;
 using Microsoft.CodeAnalysis;
 
 namespace Amazon.Lambda.Annotations.SourceGenerator.Models.Attributes
