+++ Libraries/src/Amazon.Lambda.Annotations.SourceGenerator/Models/Attributes/HttpApiAttributeBuilder.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Linq;
+using Amazon.Lambda.Annotations.APIGateway;
 using Microsoft.CodeAnalysis;
 
 namespace Amazon.Lambda.Annotations.SourceGenerator.Models.Attributes
