+++ Libraries/src/Amazon.Lambda.Annotations.SourceGenerator/Models/Attributes/TemplateParametersExtension.cs
@@ -1,3 +1,4 @@
+using Amazon.Lambda.Annotations.APIGateway;
 using System;
 using System.Collections.Generic;
 using System.Linq;
