+++ Libraries/src/Amazon.Lambda.Annotations/APIGateway/FromBodyAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Amazon.Lambda.Annotations
+namespace Amazon.Lambda.Annotations.APIGateway
 {
     [AttributeUsage(AttributeTargets.Parameter)]
     public class FromBodyAttribute : Attribute
