+++ Libraries/src/Amazon.Lambda.Annotations/APIGateway/FromHeaderAttribute.cs
@@ -1,9 +1,9 @@
 using System;
 
-namespace Amazon.Lambda.Annotations
+namespace Amazon.Lambda.Annotations.APIGateway
 {
     [AttributeUsage(AttributeTargets.Parameter)]
-    public class FromQueryAttribute : Attribute, INamedAttribute
+    public class FromHeaderAttribute : Attribute, INamedAttribute
     {
         public string Name { get; set; }
     }
